requirements.txt
agate==1.9.0
agate-dbf==0.2.2
agate-excel==0.4.1
agate-sql==0.7.0
aiobotocore==2.9.0
aiofiles==23.2.1
aiohttp==3.9.1
aioitertools==0.11.0
aiosignal==1.3.1
alembic==1.13.0
annotated-types==0.6.0
anyio==4.1.0
apache-airflow==2.8.0
apache-airflow-providers-amazon==8.15.0
apache-airflow-providers-cncf-kubernetes==7.13.0
apache-airflow-providers-common-sql==1.8.1
apache-airflow-providers-ftp==3.6.1
apache-airflow-providers-http==4.7.0
apache-airflow-providers-imap==3.4.0
apache-airflow-providers-postgres==5.10.0
apache-airflow-providers-sqlite==3.5.0
apispec==6.3.0
appnope==0.1.3
argcomplete==3.1.6
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.7.2
asn1crypto==1.5.1
asttokens==2.2.1
async-lru==2.0.4
attrs==23.1.0
awscli==1.32.2
Babel==2.13.1
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.2
bleach==6.1.0
blinker==1.7.0
boto3==1.34.2
botocore==1.34.2
Brotli==1.1.0
bs4==0.0.1
cachelib==0.9.0
cachetools==5.3.2
cattrs==23.2.3
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.7
clickclick==20.10.2
colorama==0.4.4
colorlog==4.8.0
comm==0.1.3
ConfigUpdater==3.2
connexion==2.14.2
contourpy==1.1.1
cron-descriptor==1.4.0
croniter==2.0.1
cryptography==41.0.7
csvkit==1.3.0
cycler==0.12.1
dbfread==2.0.7
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
dill==0.3.7
distlib==0.3.8
dnspython==2.4.2
docutils==0.16
email-validator==1.3.1
et-xmlfile==1.1.0
executing==1.2.0
fastjsonschema==2.19.1
filelock==3.13.1
Flask==2.2.5
Flask-AppBuilder==4.3.10
Flask-Babel==2.0.0
Flask-Caching==2.1.0
Flask-JWT-Extended==4.5.3
Flask-Limiter==3.5.0
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
FlightRadarAPI==1.3.24
fonttools==4.43.1
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.12.2
google-auth==2.26.1
google-re2==1.1
googleapis-common-protos==1.61.0
graphviz==0.20.1
grpcio==1.59.3
gunicorn==21.2.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.2
httpx==0.25.2
idna==3.4
importlib-metadata==6.11.0
importlib-resources==6.1.1
inflection==0.5.1
iniconfig==2.0.0
ipykernel==6.24.0
ipython==8.14.0
ipywidgets==8.1.2
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
json5==0.9.14
jsonpath-ng==1.6.0
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.1.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.3
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
kubernetes==23.6.0
kubernetes-asyncio==24.2.3
lab==8.0
lazy-object-proxy==1.9.0
leather==0.3.4
limits==3.7.0
line-profiler==4.1.2
linkify-it-py==2.0.2
lockfile==0.12.2
lxml==5.1.0
Mako==1.3.0
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
mistune==3.0.2
multidict==6.0.4
nbclient==0.9.0
nbconvert==7.16.0
nbformat==5.9.2
nest-asyncio==1.5.6
notebook==7.1.0
notebook_shim==0.2.4
numpy==1.25.1
oauthlib==3.2.2
olefile==0.47
openpyxl==3.1.2
opentelemetry-api==1.21.0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-exporter-otlp-proto-common==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-exporter-otlp-proto-http==1.21.0
opentelemetry-proto==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-semantic-conventions==0.42b0
ordered-set==4.1.0
overrides==7.7.0
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.1
parsedatetime==2.6
parso==0.8.3
pathspec==0.11.2
pendulum==2.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.9.1
pluggy==1.3.0
ply==3.11
prison==0.2.1
prometheus_client==0.20.0
prompt-toolkit==3.0.39
protobuf==4.25.1
psutil==5.9.5
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
Pygments==2.15.1
pyjokes==0.6.0
PyJWT==2.8.0
pyparsing==3.1.1
PyPDF2==3.0.1
pyscopg2==66.0.2
pyspark==3.5.0
pytest==7.4.3
python-daemon==3.0.1
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.0
python-json-logger==2.0.7
python-ldap==3.4.4
python-nvd3==0.15.0
python-slugify==8.0.1
pytimeparse==1.1.8
pytz==2023.3
pytzdata==2020.1
PyYAML==6.0.1
pyzmq==25.1.0
qtconsole==5.5.1
QtPy==2.4.1
redshift-connector==2.0.918
referencing==0.31.1
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rich-argparse==1.4.0
rpds-py==0.13.2
rsa==4.7.2
s3fs==0.4.2
s3transfer==0.9.0
scramp==1.4.4
Send2Trash==1.8.2
setproctitle==1.3.3
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==1.4.50
SQLAlchemy-JSONField==1.0.2
sqlalchemy-redshift==0.8.14
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
stack-data==0.6.2
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.4.0
terminado==0.18.0
text-unidecode==1.3
tinycss2==1.2.1
tornado==6.3.2
traitlets==5.9.0
txt2tags==3.9
types-python-dateutil==2.8.19.20240106
typing_extensions==4.8.0
tzdata==2023.3
uc-micro-py==1.0.2
unicodecsv==0.14.1
universal-pathlib==0.1.4
uri-template==1.3.0
urllib3==2.0.4
virtualenv==20.25.0
watchtower==3.0.1
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==2.2.3
widgetsnbextension==4.0.10
wrapt==1.16.0
WTForms==3.0.1
xlrd==2.0.1
yarl==1.9.3
zipp==3.17.0
